pyproject.toml
[project]
name = "thing"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "discord.py",
    "python-dotenv",
    "ollama",
    "aiofiles",
]
